Allow empty server extensions
Extensions may not be present in TLS 1.2 and lower, including the
two bytes indicating the extension length. This is the case for
imageshack.com.

The code previously assumed that at least the two bytes indicating
the extensions length are present, causing the bytes following to
be misinterpreted as extensions length.